[refactor] : supabase 관련 함수 위치 변경
