+++ app/api/notifications/mark-all-read/route.ts
@@ -1,7 +1,7 @@
 // app/api/notifications/mark-all-read/route.ts
 import { NextResponse } from 'next/server'
 
-import { createServerClientComponent } from '@/libs/supabase/server'
+import { createServerClientComponent } from '@/supabase/functions/server'
 
 export async function POST() {
   try {
