+++ app/api/notifications/mark-chat-read/route.ts
@@ -1,7 +1,7 @@
 // app/api/notifications/mark-chat-read/route.ts
 import { NextRequest, NextResponse } from 'next/server'
 
-import { createServerClientComponent } from '@/libs/supabase/server'
+import { createServerClientComponent } from '@/supabase/functions/server'
 
 export async function POST(request: NextRequest) {
   try {
