+++ app/dashboard/actions.ts
@@ -3,7 +3,7 @@
 import { cookies } from 'next/headers';
 import { revalidatePath } from 'next/cache';
 
-import { createActionClient } from '@/libs/supabase/server';
+import { createActionClient } from '@/supabase/functions/server';
 
 /**
  * 프로필 데이터 가져오기 서버 액션
